requirements.txt
-i https://pypi.org/simple
asgiref==3.6.0 ; python_version >= '3.7'
async-timeout==4.0.2 ; python_version >= '3.6'
django==4.1.7
django-appconf==1.0.5 ; python_version >= '3.6'
django-compressor==4.3.1
django-environ==0.9.0
django-redis==5.2.0
django-taggit==3.1.0
djongo==1.3.6
dnspython==2.3.0
markdown==3.4.1
pymongo==3.12.1
pytz==2022.7.1
rcssmin==1.1.1
redis==4.5.1 ; python_version >= '3.7'
rjsmin==1.2.1
sqlparse==0.2.4 ; python_version >= '3.5'
tzdata==2022.7 ; sys_platform == 'win32'
waitress==2.1.2
whitenoise==6.3.0
